refactor: 폴더 및 파일 구조 변경으로 인한 코드 리팩토링 - 같은 주제의 코드들을 각각의 모듈에 배치하도록 파일 구조 및 코드 변경
