--- src/utils/resetInput.js
@@ -1,6 +1,0 @@
-import { $ } from "./selector.js";
-
-export const resetInput = () => {
-  const $input = $("#user-input");
-  $input.value = "";
-};
